requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
pyaml==17.8.0
pysolr==3.6.0
PyYAML
requests==2.18.4
urllib3==1.22
